Fix: Construct new comment path without querying for post
By constructing the path for deleting a new comment using the
[comment.post, comment] syntax, we were creating an additional database query
for the comment's post. This is not at all needed, since all we really need is
the post's ID... But this is already stored with the comment record as post_id!
To take advantage of that, we now use post_comment_path(comment) and the extra
query is history.