Private message can only be sent to visible users
Profile visibility settings extend to private messages. User A can only
message user B if user A can see the profile of user B. Whether user B can see
the profile of user A does not matter.

What this means, in essence, is that users with private profiles can message
users with network-only profiles but not vice-versa.