Users can create new conversations
Users can message other users on the platform by writing a message addressed at
the recipient's username. In addition, this commit introduces a number of
improvements to the way new messages are handled: The controller will attempt
to find existing conversations between two users and attempt to create a new
conversation if the current one was just deleted.