Split default scopes into custom scopes
To avoid eager loading when not necessary (for example, when destroying an
object), we must split up the eager loading from the model's default scope.
In addition, several people warn strongly against using default_scope at all.
It may be necessary to get rid of it completely in the future.

See:
- http://stackoverflow.com/questions/1834159/overriding-a-rails-default-scope
- http://rails-bestpractices.com/posts/2013/06/15/default_scope-is-evil/
- http://stackoverflow.com/questions/4859099/using-scope-on-an-association