Remove secrets from .travis.yml
There is no need to store the actual values for the Mailjet API and the Mailjet
secret key. The tests only ensure that the variables are defined, so we can
simply declare them as 'spoof'. This is a) safer and b) makes it easier for
those who are creating forks.