Use poltergeist instead of webkit for capybara
From: http://stackoverflow.com/a/24108439/6451879

Poltergeist is the capybara driver for PhantomJS, a headless browser which is
built on WebKit. capybara-webkit is a capybara driver which uses WebKit
directly.

Poltergeist/PhantomJS has some big advantages over capybara-webkit:

- In my experience poltergeist/PhantomJS always times out when it should,
whereas capybara-webkit sometimes hangs.
- Its error messages are much clearer. For example, it will actually tell you if
it can't click an element that is on the page because there's another element in
front of it.
- It can be told to re-raise Javascript errors as test errors (by instantiating
the driver with js_errors: true).
- PhantomJS is much easier to install than standalone WebKit. PhantomJS provides
a nearly dependency-free executable that you can download, while standalone
WebKit has many OS library dependencies which you may have to upgrade or
otherwise fiddle with.