Fix: Database Cleaner to avoid false errors on Capybara JS specs
We were encountering false errors on Capybara feature specs that use Javascript
(:js => true). Records in the database would suddenly not be able to be found.
Not quite clear why this is the case, but this patch seems to fix the problem.

Credits go to kugaevsky@Coderwall: https://coderwall.com/p/ahtb7w/